+++ example/lib/screens/dialpad_basic.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dialpad/flutter_dialpad.dart';
 import 'package:flutter_dialpad_example/widgets/ui_app_bar.dart';
 
-class DialPadPlain extends StatelessWidget {
-  const DialPadPlain({super.key});
+class DialPadBasic extends StatelessWidget {
+  const DialPadBasic({super.key});
 
   void _makeCall(String number) {
     print('calling $number');
@@ -16,10 +16,12 @@ class DialPadPlain extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: UIAppBar(title: 'DialPad Plain-style'),
+      appBar: UIAppBar(title: 'DialPad Basic-style'),
       body: DialPad(
         makeCall: _makeCall,
         keyPressed: _keyPressed,
+        scalingType: ScalingType.fixed,
+        scalingSize: ScalingSize.small,
       ),
     );
   }