+++ lib/src/widgets/buttons/button_type.dart
@@ -1,4 +1,0 @@
-enum ButtonType {
-  rectangle,
-  circle;
-}