+++ lib/src/widgets/phone_text_field.dart
@@ -1,12 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-class PhoneTextField extends StatelessWidget {
+import '../mixins/scalable.dart';
+import 'scalable/scalable.dart';
+
+typedef CopyToClipboardBuilder = Widget Function(BuildContext context, VoidCallback onCopyPressed);
+
+class PhoneTextField extends StatelessWidget with Scalable {
   /// TextStyle for the text field.
   final TextStyle? textStyle;
 
   /// The background color of the text field. Defaults to [Colors.white].
   final Color color;
+
+  /// The text color of the text field. Defaults to [Colors.grey].
+  final Color textColor;
+
+  /// Font size for the text field, as a percentage of the screen height. Defaults to 25.
+  final double textSize;
 
   /// The decoration to show around the text field. Defaults to [InputDecoration(border: InputBorder.none)].
   final InputDecoration decoration;
@@ -26,26 +37,52 @@ class PhoneTextField extends StatelessWidget {
   /// The alignment of the text field. Defaults to [TextAlign.center].
   final TextAlign textAlign;
 
+  /// Builder for the copyToClipboard widget. Defaults to [_defaultCopyToClipboardBuilder].
+  final CopyToClipboardBuilder? copyToClipboardBuilder;
+
+  /// [ScalingType] for the button. Defaults to [ScalingType.fixed].
+  final ScalingType scalingType;
+
+  /// [ScalingSize] for the button. Defaults to [ScalingSize.small].
+  final ScalingSize scalingSize;
+
   const PhoneTextField({
     super.key,
     this.textStyle,
     this.color = Colors.white,
+    this.textColor = Colors.grey,
+    this.textSize = 15,
     this.decoration = const InputDecoration(border: InputBorder.none),
     this.onChanged,
     this.readOnly = false,
     this.textAlign = TextAlign.center,
     this.copyToClipboard = false,
     required this.controller,
+    this.copyToClipboardBuilder,
+    this.scalingType = ScalingType.fixed,
+    this.scalingSize = ScalingSize.small,
   });
 
   void _onCopyPressed() {
     Clipboard.setData(ClipboardData(text: controller.text));
   }
 
+  Widget _defaultCopyToClipboardBuilder() {
+    return IconButton(icon: Icon(Icons.copy), onPressed: _onCopyPressed);
+  }
+
   @override
   Widget build(BuildContext context) {
+    final screenSize = MediaQuery.of(context).size;
+    final size = rescale(screenSize, scalingType, textSize, scalingSize: scalingSize);
+
+    final _builtTextStyle = TextStyle(
+      color: textColor,
+      fontSize: size,
+    );
+
     final textField = TextFormField(
-      style: textStyle,
+      style: textStyle ?? _builtTextStyle,
       decoration: decoration,
       readOnly: readOnly,
       textAlign: textAlign,
@@ -57,7 +94,7 @@ class PhoneTextField extends StatelessWidget {
       return Row(
         children: [
           Expanded(child: textField),
-          IconButton(icon: Icon(Icons.copy), onPressed: _onCopyPressed),
+          copyToClipboardBuilder?.call(context, _onCopyPressed) ?? _defaultCopyToClipboardBuilder(),
         ],
       );
     } else {