+++ lib/src/widgets/scalable/button_type.dart
@@ -1,0 +1,8 @@
+/// [ButtonType] is an enum that defines the shape of a button
+enum ButtonType {
+  /// [ButtonType.rectangle] will clip the button to a rectangle e.g. an Windows/metro design
+  rectangle,
+
+  /// [ButtonType.circle] will clip the button to a circle e.g. an iPhone keypad
+  circle;
+}