+++ lib/src/widgets/widgets.dart
@@ -2,3 +2,4 @@ export 'buttons/buttons.dart';
 export 'keypad_grid.dart';
 export 'phone_text_field.dart';
 export 'keypad_focus_node.dart';
+export 'scalable/scalable.dart';